[#5] Organize QoS files
