+++ rmw_iceoryx2_cxx/src/impl/runtime/publisher.cpp
@@ -13,9 +13,9 @@
 #include "iox2/bb/slice.hpp"
 #include "rmw_iceoryx2_cxx/impl/common/error_message.hpp"
 #include "rmw_iceoryx2_cxx/impl/common/names.hpp"
-#include "rmw_iceoryx2_cxx/impl/common/qos_attributes.hpp"
 #include "rmw_iceoryx2_cxx/impl/message/introspection.hpp"
 #include "rmw_iceoryx2_cxx/impl/middleware/iceoryx2.hpp"
+#include "rmw_iceoryx2_cxx/impl/qos/attributes.hpp"
 
 namespace rmw::iox2
 {
