+++ rmw_iceoryx2_cxx/src/impl/runtime/subscriber.cpp
@@ -12,8 +12,8 @@
 #include "iox2/bb/into.hpp"
 #include "rmw_iceoryx2_cxx/impl/common/error_message.hpp"
 #include "rmw_iceoryx2_cxx/impl/common/names.hpp"
-#include "rmw_iceoryx2_cxx/impl/common/qos_attributes.hpp"
 #include "rmw_iceoryx2_cxx/impl/middleware/iceoryx2.hpp"
+#include "rmw_iceoryx2_cxx/impl/qos/attributes.hpp"
 
 namespace rmw::iox2
 {
