+++ rmw_iceoryx2_cxx/test/test_impl_publisher.cpp
@@ -12,8 +12,8 @@
 #include "iox2/bb/optional.hpp"
 #include "rmw/qos_profiles.h"
 #include "rmw_iceoryx2_cxx/impl/common/create.hpp"
-#include "rmw_iceoryx2_cxx/impl/common/qos.hpp"
-#include "rmw_iceoryx2_cxx/impl/common/qos_attributes.hpp"
+#include "rmw_iceoryx2_cxx/impl/qos/attributes.hpp"
+#include "rmw_iceoryx2_cxx/impl/qos/qos.hpp"
 #include "rmw_iceoryx2_cxx/impl/runtime/context.hpp"
 #include "rmw_iceoryx2_cxx/impl/runtime/publisher.hpp"
 #include "rmw_iceoryx2_cxx_test_msgs/msg/defaults.hpp"
